# specs/002-roles-permissions-system/openapi.yaml
openapi: 3.0.3
info:
  title: WasteFlow Roles & Permissions API
  description: |
    Comprehensive roles and permissions management system for WasteFlow application.

    This API supports:
    - Multi-tenant role-based access control (RBAC)
    - Attribute-based access control (ABAC) policies
    - Consultant multi-tenant context switching
    - Temporary permission grants with auto-expiration
    - Comprehensive audit logging for regulatory compliance
    - Custom role creation for enterprise clients
    - Permission request workflows

    ## Authentication
    All endpoints require JWT Bearer authentication obtained via SPID/CIE login.
    The JWT contains user identity, tenant context, and SPID fiscal code for audit traceability.

    ## Multi-Tenancy
    - Schema-per-tenant architecture with Row-Level Security (RLS)
    - Tenant context extracted from JWT claims
    - Consultants can switch tenant context via dedicated endpoints
    - All operations scoped to current tenant unless explicitly cross-tenant (consultants only)

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute per user
    - Permission check endpoints: 1000 requests/minute per user (high-frequency caching)
    - Audit log queries: 20 requests/minute per user (expensive operations)
    - Bulk operations: 10 requests/minute per user

    ## Compliance
    - All audit logs retained for 10+ years per D.Lgs. 152/2006
    - Immutable audit trail with cryptographic integrity
...

  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /roles/{roleId}:
    get:
      tags: [Roles]
      summary: Get role details
      description: |
        Returns complete role information including:
        - Permission matrix
        - Number of assigned users
        - Creation/modification history
        - Preview of capabilities
      operationId: getRoleById
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
      responses:
        '200':
          description: Role details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 100/minute

    patch:
      tags: [Custom Roles]
      summary: Update custom role
      description: |
        Updates custom role permissions or metadata.
        System roles cannot be modified.
        Permission changes affect all users with this role immediately.
        Audit log entry created with before/after snapshots.
      operationId: updateCustomRole
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - cannot modify system roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 10/minute

    delete:
      tags: [Custom Roles]
      summary: Delete custom role
      description: |
        Deletes custom role if no users are currently assigned.
        System roles cannot be deleted.
        If users are assigned, returns error with reassignment guidance.
      operationId: deleteCustomRole
      parameters:
        - $ref: '#/components/parameters/RoleIdParam'
      responses:
        '204':
          description: Role deleted successfully
        '400':
          description: Cannot delete - users still assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: ROLE_IN_USE
                message: This role is assigned to 12 users. Reassign them first before deleting.
                statusCode: 400
                details:
                  assignedUserCount: 12
                  assignedUserIds: ["user-1", "user-2", "..."]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 10/minute
...
  /users/{userId}/roles:
    get:
      tags: [Roles]
      summary: Get user's assigned roles
      description: |
        Returns all roles assigned to user including:
        - Active roles
        - Temporary role elevations
        - Facility-scoped roles
        - Expiration timestamps
      operationId: getUserRoles
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - name: includeExpired
          in: query
          schema:
            type: boolean
            default: false
          description: Include expired role assignments in results
      responses:
        '200':
          description: User roles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 100/minute

    post:
      tags: [Roles]
      summary: Assign role to user
      description: |
        Assigns one or more roles to user within current tenant.
        Supports facility-scoped roles for large enterprises.
        Creates audit log entry with assignment justification.
        Invalidates user's permission cache immediately.
      operationId: assignRoleToUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
      responses:
        '201':
          description: Role assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleAssignmentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Conflict - role already assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-rate-limit: 10/minute
  /users/{userId}/roles/{roleId}:
    delete:
      tags: [Roles]
      summary: Revoke role from user
      description: |
        Removes role assignment from user.
        Prevents removal if this would leave tenant with no administrators.
        Creates audit log entry with revocation reason.
        Invalidates user's permission cache immediately.
      operationId: revokeRoleFromUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/RoleIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  minLength: 10
                  maxLength: 500
                  description: Reason for revocation (mandatory for audit)
              required: [reason]
      responses:
        '204':
          description: Role revoked successfully
        '400':
          description: Cannot revoke - would leave tenant without admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: LAST_ADMIN_PROTECTION
                message: At least one administrator is required per organization
                statusCode: 400
                details:
                  currentAdminCount: 1
                  requiredMinimum: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 10/minute
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via SPID/CIE authentication.
        Include in Authorization header: `Bearer <token>`

        JWT contains claims:
        - userId: User unique identifier
        - tenantId: Current tenant context
        - fiscalCode: SPID fiscal code for audit
        - roles: User's role IDs
        - permissions: Cached permission set
        - exp: Token expiration timestamp
  parameters:
...
    RoleIdParam:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Role unique identifier
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User unique identifier
...
  schemas:
...
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Role unique identifier
        name:
          type: string
          description: Role name (unique per tenant)
          example: Site Manager Milano
        description:
          type: string
          description: Human-readable role description
          example: Manages waste operations for Milano facility
        isSystemRole:
          type: boolean
          description: System role (cannot be modified/deleted)
        permissions:
          type: array
          items:
            type: string
          description: Array of permission strings (format resource:action:scope)
          example: ["fir:create:facility", "fir:read:all", "registry:write:facility"]
        assignedUserCount:
          type: integer
          description: Number of users currently with this role
        facilityScoped:
          type: boolean
          description: Whether role can be facility-scoped
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
          description: User who created this custom role
      required: [id, name, description, isSystemRole, permissions, assignedUserCount, createdAt]
    RoleDetailResponse:
      allOf:
        - $ref: '#/components/schemas/RoleResponse'
        - type: object
          properties:
            permissionsByModule:
              type: object
              description: Permissions organized by module for clear display
              additionalProperties:
                type: object
                properties:
                  module:
                    type: string
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/PermissionDetail'
              example:
                fir:
                  module: Formulari Rifiuti (FIR)
                  permissions:
                    - resource: fir
                      action: create
                      scope: facility
                      description: Create new FIR for assigned facilities
                registry:
                  module: Registry Management
                  permissions:
                    - resource: registry
                      action: read
                      scope: all
                      description: View all registry entries
            assignedUsers:
              type: array
              items:
                $ref: '#/components/schemas/UserRoleSummary'
              description: Summary of users with this role (first 10)
            changeHistory:
              type: array
              items:
                $ref: '#/components/schemas/RoleChangeRecord'
              description: Recent modification history (last 10 changes)
    PermissionDetail:
      type: object
      properties:
        resource:
          type: string
          description: Resource type
          example: fir
        action:
          type: string
          description: Action on resource
          enum: [create, read, update, delete, approve, sign, export]
        scope:
          type: string
          description: Permission scope
          enum: [own, facility, all]
        description:
          type: string
          description: Human-readable permission description (Italian)
        sensitive:
          type: boolean
          description: High-risk permission requiring extra scrutiny
      required: [resource, action, scope, description]
    UserRoleSummary:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        facilityIds:
          type: array
          items:
            type: string
            format: uuid
          description: Facility scope (if applicable)
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
          nullable: true
    RoleChangeRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        changeType:
          type: string
          enum: [CREATE, UPDATE, DELETE, ASSIGN, REVOKE]
        changedBy:
          type: string
          format: uuid
        changedByName:
          type: string
        reason:
          type: string
        timestamp:
          type: string
          format: date-time
        beforeSnapshot:
          type: object
          description: State before change (JSON)
        afterSnapshot:
          type: object
          description: State after change (JSON)
...
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          minLength: 10
          maxLength: 500
        permissions:
          type: array
          items:
            type: string
        reason:
          type: string
          minLength: 10
          maxLength: 500
          description: Reason for modification (mandatory for audit)
      required: [reason]
...
    UserRolesResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRoleAssignment'
        effectivePermissions:
          type: array
          items:
            type: string
          description: Union of all permissions from assigned roles
        temporaryElevations:
          type: array
          items:
            $ref: '#/components/schemas/TemporaryGrantSummary'
          description: Active temporary permission grants
      required: [userId, roles, effectivePermissions]
    UserRoleAssignment:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        roleDescription:
          type: string
        facilityIds:
          type: array
          items:
            type: string
            format: uuid
          description: Facility scope (empty array = all facilities)
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: string
          format: uuid
        assignedByName:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Role expiration (null = permanent)
        isDelegated:
          type: boolean
          description: Whether this is delegated role from another user
        delegationReason:
          type: string
          nullable: true
      required: [roleId, roleName, roleDescription, assignedAt, assignedBy]
    AssignRoleRequest:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
          description: Role to assign
        facilityIds:
          type: array
          items:
            type: string
            format: uuid
          description: Optional facility scope restriction
        expiresAt:
          type: string
          format: date-time
          description: Optional expiration timestamp
        reason:
          type: string
          minLength: 10
          maxLength: 500
          description: Reason for assignment (mandatory for audit)
      required: [roleId, reason]
    UserRoleAssignmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Assignment unique identifier
        userId:
          type: string
          format: uuid
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        facilityIds:
          type: array
          items:
            type: string
            format: uuid
        assignedAt:
          type: string
          format: date-time
        assignedBy:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
          nullable: true
        permissionsInvalidated:
          type: boolean
          description: Whether user's permission cache was invalidated
      required: [id, userId, roleId, roleName, assignedAt, assignedBy]
...
    TemporaryGrantSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        permissions:
          type: array
          items:
            type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        remainingMinutes:
          type: integer
          description: Minutes until expiration
        justification:
          type: string
      required: [id, permissions, startTime, endTime]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: PERMISSION_DENIED
        message:
          type: string
          description: Human-readable error message (Italian)
          example: Non hai il permesso necessario per eseguire questa azione
        statusCode:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint that generated error
      required: [error, message, statusCode, timestamp]
    ValidationErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
        message:
          type: string
          description: Validation error message
        value:
          description: Invalid value provided
        constraint:
          type: string
          description: Validation constraint violated
      required: [field, message]
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Authentication required. Please login with SPID/CIE.
            statusCode: 401
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
    ForbiddenError:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Non hai il permesso necessario per eseguire questa azione
            statusCode: 403
            details:
              requiredPermission: roles:create
              currentRole: OPERATOR
              contactAdministrator: admin@example.com
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Role non trovato
            statusCode: 404
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles/123e4567-e89b-12d3-a456-426614174000
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  validationErrors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed
            statusCode: 400
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
            validationErrors:
              - field: name
                message: Role name must be at least 3 characters
                value: ab
                constraint: minLength
              - field: permissions
                message: At least one permission is required
                constraint: minItems
...